+++ packages/react-magma-dom/src/i18n/interface.ts
@@ -236,4 +236,14 @@ export interface I18nInterface {
     amSelectedAnnounce: string;
     pmSelectedAnnounce: string;
   };
+  wizard: {
+    navigationLabel: string;
+    optional: string;
+    actions: {
+      previous: string;
+      next: string;
+      submit: string;
+      cancel: string;
+    };
+  };
 }