+++ packages/schema-renderer/.eslintrc.js
@@ -1,0 +1,10 @@
+const eslintrc = require('../../.eslintrc');
+
+module.exports = {
+  ...eslintrc,
+  rules: {
+    ...eslintrc.rules,
+    complexity: 0,
+    'no-unused-vars': 0,
+  },
+};